store_electricity_price_from_jpex/requirements.txt
# Function dependencies, for example:
# package>=version
datetime==4.3
pytz==2022.7.1
requests==2.26.0
polars==0.17.3
google-cloud-firestore==2.11.0
msgpack==1.0.5
google-cloud-storage==2.9.0